Refactoring to register multiple sources of same type
Refactored the code so that multiple sources of same type can be registered.
The FileBasedMarkdownSource is no longer a spring managed bean, it's instance need to be created explicitly.
The sources are handled by MarkdownSourceProvider.
Refactored the installer, added AbstractInstaller to ensure that HelpDocs are installed as per installer config.